+++ bubulle-py/utils/file_utils.py
@@ -25,7 +25,10 @@
 # SOFTWARE.#
 import os
 
-read = lambda n: open(n, 'r').read()
+
+def read_file(path):
+    with open(path, 'r', encoding="utf-8") as f:
+        return f.read()
 
 
 def get_path(args):
@@ -36,11 +39,9 @@ def get_path(args):
     return path
 
 
-def is_tempfile(path):
+def is_temp_file(path):
     if path.endswith(".tmp"):
-        try:
+        if os.path.isfile(path):
             os.remove(path)
-        except:
-            pass
         return 1
     return 0
